Remove lucene-spatial
It's going to be removed in Lucene 8.5
See https://issues.apache.org/jira/browse/LUCENE-8369

ES replaced it in a different way, see
https://github.com/elastic/elasticsearch/commit/a8870ef98c9a25961914b408f25d27dabd3a5a11

+++ es/es-server/src/main/java/org/elasticsearch/common/geo/GeoHashUtils.java
@@ -19,7 +19,6 @@
 package org.elasticsearch.common.geo;
 
 import org.apache.lucene.geo.Rectangle;
-import org.apache.lucene.spatial.util.MortonEncoder;
 import org.apache.lucene.util.BitUtil;
 
 import java.util.Collection;
@@ -71,7 +70,7 @@ public class GeoHashUtils {
         long b;
         long l = 0L;
         for(char c : hash.toCharArray()) {
-            b = (long)(BASE_32_STRING.indexOf(c));
+            b = BASE_32_STRING.indexOf(c);
             l |= (b<<(level--*5));
             if (level < 0) {
                 // We cannot handle more than 12 levels
